refactor(nav): Refactor nav off main.dart
